+++ pycons3rt3/cons3rtconfig.py
@@ -20,10 +20,9 @@ mod_logger = Logify.get_name() + '.cons3rtconfig'
 site_urls = {
     'arcusgov': 'https://app.arcus-cloud.io/rest/api/',
     'arcusmil': 'https://app.arcus.mil/rest/api/',
-    'ci': 'https://api.ci.cons3rt.io/rest/api/',
-    'ci2': 'https://api.ci2.cons3rt.io/rest/api/',
-    'dev': 'https://api.dev.cons3rt.io/rest/api/',
-    'qa': 'https://api.qa.cons3rt.io/rest/api/',
+    'ci': 'https://app.ci.cons3rt.io/rest/api/',
+    'dev': 'https://app.dev.cons3rt.io/rest/api/',
+    'qa': 'https://app.qa.cons3rt.io/rest/api/',
 }
 
 # The default site selection
@@ -32,7 +31,8 @@ default_api_url = site_urls['arcusmil']
 # List of sites that require certificate-based auth
 cert_auth_sites = [
     site_urls['arcusgov'],
-    site_urls['arcusmil']
+    site_urls['arcusmil'],
+    site_urls['qa']
 ]
 
 # String representation of the list of sites