+++ pycons3rt3/pycons3rtlibs.py
@@ -76,7 +76,7 @@ class HostActionResult(object):
 class RestUser(object):
 
     def __init__(self, rest_api_url, token, project=None, cert_file_path=None, key_file_path=None, username=None,
-                 cert_bundle=None):
+                 cert_bundle=None, site_default_project=False):
         self.rest_api_url = rest_api_url
         self.token = token
         self.project_name = project
@@ -84,12 +84,19 @@ class RestUser(object):
         self.key_file_path = key_file_path
         self.username = username
         self.cert_bundle = cert_bundle
+        self.site_default_project = site_default_project
 
     def __str__(self):
-        base_str = 'CONS3RT user for site [{s}] in project: {p}'.format(s=self.rest_api_url, p=self.project_name)
+        user_str = 'CONS3RT user for site [{s}] in project [{p}]'.format(s=self.rest_api_url, p=self.project_name)
         if self.cert_file_path:
-            return base_str + ', using cert auth: {c}'.format(c=self.cert_file_path)
+            user_str += ', using cert [{c}]'.format(c=self.cert_file_path)
+            if self.key_file_path:
+                user_str += ' and key [{k}]'.format(k=self.key_file_path)
         elif self.username:
-            return base_str + ', using username auth: {u}'.format(u=self.username)
-        else:
-            return base_str
+            user_str += ', using username [{u}]'.format(u=self.username)
+        if self.cert_bundle:
+            if isinstance(self.cert_bundle, str):
+                user_str += ', using CA cert bundle [{b}]'.format(b=self.cert_bundle)
+        if self.site_default_project:
+            user_str += ' (default)'
+        return user_str